move model mgmt helper funcs
